Reorganize files by purpose
