+++ src/ToolboxClient.ts
@@ -1,7 +1,7 @@
 import {RedditAPIClient, WikiPage} from '@devvit/public-api';
-import {Usernote, UsernoteInit} from '../types/Usernote';
-import {SubredditConfig} from './SubredditConfig';
-import {Usernotes} from './Usernotes';
+import {SubredditConfig} from './config/SubredditConfig';
+import {Usernote, UsernoteInit} from './usernotes/Usernote';
+import {Usernotes} from './usernotes/Usernotes';
 
 /** The name of the wiki page where Toolbox stores usernotes. */
 const TB_USERNOTES_PAGE = 'usernotes';
