+++ src/config/SubredditConfig.ts
@@ -2,11 +2,11 @@ import {
 	DEFAULT_CONFIG,
 	DEFAULT_USERNOTE_TYPES,
 	migrateConfigToLatestSchema,
-} from '../helpers/config';
-import {RawSubredditConfig, RawUsernoteType} from '../types/RawSubredditConfig';
+} from './config';
+import {RawSubredditConfig, RawUsernoteType} from './RawSubredditConfig';
 
 // type imports for doc references
-import type {Usernote} from '../types/Usernote';
+import type {Usernote} from '../usernotes/Usernote';
 
 /**
  * A class that interfaces with the raw contents of a subreddit's `toolbox`
